Update docs and comments. Also, move MultipleDistinctRewriter to the analysis package.
