+++ sql/core/src/main/scala/org/apache/spark/sql/execution/SparkStrategies.scala
@@ -174,10 +174,9 @@ private[sql] abstract class SparkStrategies extends QueryPlanner[SparkPlan] {
           aggregateExpressions.partition(_.isDistinct)
         if (functionsWithDistinct.map(_.aggregateFunction.children).distinct.length > 1) {
           // This is a sanity check. We should not reach here when we have multiple distinct
-          // column sets (aggregate.NewAggregation will not match).
-          sys.error(
-            "Multiple distinct column sets are not supported by the new aggregation" +
-              "code path.")
+          // column sets. Our MultipleDistinctRewriter should take care this case.
+          sys.error("You hit a query analyzer bug. Please report your query to " +
+            "Spark user mailing list.")
         }
 
         val namedGroupingExpressions = groupingExpressions.map {
